Introduces validation of the user id of a request, if it is the same as the email of the user.

+++ Pantrymony.back/Lambda/Http/DeleteVictual.cs
@@ -21,6 +21,10 @@ public class DeleteVictual
             AWSSDKHandler.RegisterXRayForAllServices();
             var victualId = request.QueryStringParameters[Constants.VictualIdTag];
             var userId = request.QueryStringParameters[Constants.UserIdTag];
+            if (!request.WasSentByUser(userId))
+            {
+                return HttpStatusCode.Unauthorized.AsApiGatewayProxyResponse();
+            }
             await UserVictualsService.DeleteUserVictualAsync(userId, victualId, context.GetCustomLogger());
         }
         catch (Exception e)
